requirements.txt
fastapi
uvicorn
pymupdf
google-genai
requests
pydantic